#102 Fix lint errors and migrate to TS
